# docs/openapi-store-members.yaml
openapi: 3.1.0
info:
  title: Store Members API
  version: 1.0.0
  description: >-
    API de gestion des membres de magasin (Store Members).
  contact:
    name: Backend Team
    email: dev@company.tld
servers:
  - url: http://localhost:8000/api/v1
    description: Local
security:
  - bearerAuth: []
paths:
  /stores/{store}/members:
    get:
      summary: Liste paginée des membres d'un magasin
      operationId: listStoreMembers
      tags: [Store Members]
      parameters:
        - $ref: '#/components/parameters/StoreId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Liste paginée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedStoreMemberList'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      summary: Création d'un membre
      operationId: createStoreMember
      tags: [Store Members]
      parameters:
        - $ref: '#/components/parameters/StoreId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreMemberCreateRequest'
      responses:
        '201':
          description: Créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreMemberSingleResponse'
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /stores/{store}/members/{store_member}:
    get:
      summary: Détails d'un membre
      operationId: getStoreMember
      tags: [Store Members]
      parameters:
        - $ref: '#/components/parameters/StoreId'
        - $ref: '#/components/parameters/StoreMemberId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreMemberSingleResponse'
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      summary: Mise à jour partielle
      operationId: updateStoreMember
      tags: [Store Members]
      parameters:
        - $ref: '#/components/parameters/StoreId'
        - $ref: '#/components/parameters/StoreMemberId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreMemberUpdateRequest'
      responses:
        '201':
          description: Mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreMemberSingleResponse'
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: Suppression logique d'un membre
      operationId: deleteStoreMember
      tags: [Store Members]
      parameters:
        - $ref: '#/components/parameters/StoreId'
        - $ref: '#/components/parameters/StoreMemberId'
      responses:
        '204': { description: Supprimé }
        '409':
          description: Conflit (ex. shift actif)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    StoreId:
      in: path
      name: store
      required: true
      schema: { type: integer, minimum: 1 }
    StoreMemberId:
      in: path
      name: store_member
      required: true
      schema: { type: integer, minimum: 1 }
...
    Page:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    PerPage:
      in: query
      name: per_page
      schema: { type: integer, minimum: 5, maximum: 50, default: 15 }
  responses:
    Unauthorized:
      description: Non authentifié
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Accès refusé
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ValidationError:
      description: Erreur de validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationErrorResponse' }
    ServerError:
      description: Erreur interne
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    StoreMemberRole:
      type: string
      enum: [owner, manager, cashier, kitchen]
    StoreMember:
      type: object
      properties:
        id: { type: integer }
        store_id: { type: integer }
        name: { type: string }
        role: { $ref: '#/components/schemas/StoreMemberRole' }
        code_number: { type: integer, description: 'Numéro interne' }
        code: { type: string, description: 'Code formaté (ex: MGR-003)' }
        is_active: { type: boolean }
        meta: { type: object, additionalProperties: true }
        permissions:
          type: array
          items: { type: string }
        last_login_at: { type: string, format: date-time, nullable: true }
        login_count: { type: integer }
        failed_attempts: { type: integer }
        locked_until: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    StoreMemberCreateRequest:
      type: object
      required: [name, role]
      properties:
        name: { type: string, maxLength: 120 }
        role: { $ref: '#/components/schemas/StoreMemberRole' }
        pin: { type: string, pattern: '^\\d{4,8}$', description: 'Obligatoire pour owner/manager' }
        is_active: { type: boolean, default: true }
        meta: { type: object, additionalProperties: true }
    StoreMemberUpdateRequest:
      type: object
      properties:
        name: { type: string, maxLength: 120 }
        role: { $ref: '#/components/schemas/StoreMemberRole' }
        pin: { type: string, pattern: '^\\d{4,8}$' }
        is_active: { type: boolean }
        meta: { type: object, additionalProperties: true }
        permissions:
          type: array
          items: { type: string }
...
    StoreMemberSingleResponse:
      type: object
      properties:
        message: { type: string }
        data: { $ref: '#/components/schemas/StoreMember' }
    PaginatedStoreMemberList:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/StoreMember' }
        meta: { $ref: '#/components/schemas/PaginationMeta' }
        links: { $ref: '#/components/schemas/PaginationLinks' }
    PaginationMeta:
      type: object
      properties:
        current_page: { type: integer }
        from: { type: integer, nullable: true }
        last_page: { type: integer }
        path: { type: string }
        per_page: { type: integer }
        to: { type: integer, nullable: true }
        total: { type: integer }
    PaginationLinks:
      type: object
      properties:
        first: { type: string, nullable: true }
        last: { type: string, nullable: true }
        prev: { type: string, nullable: true }
        next: { type: string, nullable: true }
...
    ErrorResponse:
      type: object
      properties:
        message: { type: string }
        code: { type: string, nullable: true }
        errors:
          type: object
          additionalProperties: true
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items: { type: string }
...